+++ dataactvalidator/validation_handlers/validationManager.py
@@ -22,7 +22,7 @@ from dataactcore.utils.responseException import ResponseException
 from dataactcore.utils.jsonResponse import JsonResponse
 from dataactcore.utils.report import get_cross_file_pairs, report_file_name
 from dataactcore.utils.statusCode import StatusCode
-from dataactcore.aws.s3UrlHandler import S3Handler
+from dataactcore.aws.s3Handler import S3Handler
 from dataactvalidator.filestreaming.csvS3Reader import CsvS3Reader
 from dataactvalidator.filestreaming.csvLocalReader import CsvLocalReader
 from dataactvalidator.filestreaming.csvLocalWriter import CsvLocalWriter
